Verify if transaction exists

+++ src/services/DeleteTransactionService.ts
@@ -1,7 +1,7 @@
 import { getCustomRepository } from 'typeorm';
-// import AppError from '../errors/AppError';
 
 import TransactionsRepository from '../repositories/TransactionsRepository';
+import AppError from '../errors/AppError';
 
 interface RequestDTO {
   id: string;
@@ -10,6 +10,12 @@ interface RequestDTO {
 class DeleteTransactionService {
   public async execute({ id }: RequestDTO): Promise<void> {
     const transactionsRepository = getCustomRepository(TransactionsRepository);
+
+    const transaction = await transactionsRepository.findOne(id);
+
+    if (!transaction) {
+      throw new AppError('Transaction not exists');
+    }
 
     await transactionsRepository.delete(id);
   }
